LU-0000 klnds: convert list_for_each to list_for_each_entry
Nearly all uses of list_for_each() in the various klnds are
immediately followed by a list_entry() call, and this is the only use
of the loop iterator.  So we can convert these to
list_for_each_entry() and discard the loop iterator.

One case which isn't quite so straight forward is in
ksocknal_push_peer() where the variable used as the
new loop iterator it tested after the loop.

This is no longer safe as list_for_each_entry() can
leave it pointing at an undefined value.  In this
case we can easily test the 'i' variable to get the
same effect.

Note that we don't change list-walking for the knsd_peers
hash table.  This will be changed differently in a separate patch.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I910ee701bf13c8d63abc115a0f8846d03e40ded6