LU-0000 lnet: convert list_for_each to list_for_each_entry
Most uses for list_for_each() in lnet after followed promptly by a
list_entry() call on the loop variable.  This can be changed to
list_for_each_entry() with the loop variable discared, and sometime
the declaration of the new loop variable move outside the loop.

In lnet_parse_route() we need to introduce a new loop variable 'ltb2'
as 'ltb' was used for two nested loops.  Also, we cannot remove the
loop variables tmp1 and tmp2 as they are use in the same function for
a separate purpose.

Also re-indent some lines and join up a split string.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: Ie73f411a448c15028a239e71eccbff05fd9edd4a