LU-0000 ptlrpc: convert list_for_each to list_for_each_entry
All uses of list_for_each() in ptlrpc is followed immediately by a
list_entry() call on the loop variable.  These can be changed to
list_for_each_entry() and the loop variable can mostly be discarded.

In a couple of places, the declaration of the new loop variable need
to be move out of the loop.  In one case - ptlrpc_set_wait() - the old
loop variable cannot be discarded as it is needed for another loop.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: If216f0cd2bbd2a676f8dd305b3f6cd2eea40e6ec