LU-0000 ldlm: convert list_for_each to list_for_each_entry
All uses of list_for_each() in lnet are followed promptly by a
list_entry() call on the loop variable, and they can be converted to
list_for_each_entry().  Some of these conversions are straight
forward, but others are more complex.

The list that ldlm_process_flock_lock() walks groups together entries
with the same lock owner.  So we need to walk through the list until
we find a lock with the lock owner that we are interested in, or until
we get to the end of the list (because the lock owner is not yet
present).  That place is where the new lock might be inserted.

Current code uses the variable 'ownlocks' to track the list_head where
insertion will happen.  With this patch, the single loop variable
'lock' is used to track this location, though it is possible that at
the time of insertion, it won't actually point to a valid 'ldlm_lock',
but lock->l_res_link will still be the correct list_head for insertion
(the head of the whole list).

In some cases, the code first walks the list look for locks from
*other* lock owners that might conflict, and then walks the list
looking only at locks for the same lock-owner.  The old code would
record the location of the same-lock-owners in that first pass.  In
the new code, this optimization is skipped, as the code is cleaner
that way.

ldlm_inode_bits_compat_queue() has similar complexity, but here the
list has grouped by mode, and eahcmode is grouped by policy.  We need
to keep track of the end of the current mode, and of the current
policy, so that we can keep track of where in the list we are.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I9018a1a84c37da5afa36b3bfd709654eba668e4b