LU-0000 lustre: convert remaining list_for_each() calls.
Most remaining list_for_each() calls are followed by
a list_entry() call on the loop variable, and they can
be converted to list_for_each_entry().

Most of these conversions are quite straight forward.  Possibly the
most interesting is in mgs_barrier_glimpse_lock() where we need
to test if lock is at the end of the list.  Rather than
testing if ->next points to the head of the list, we test
if lock is list_last_entry() of the list.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I25b3e74a3eb44ea17b889042d3423e6013da2f5c