LU-0000 lnet: don't embed whole lnet_md in lnet_event
Of the several fields that are copied into the lnet_md which is
embedded in an lnet_event, only three are ever used:
  start, user_ptr, md_options

Rather than allocating space for the whole structure, copying only
some in, and then using even fewer; just store the fields that are
needed.

Note that a comment suggests that the "threshold" value is
interesting, but this value is never used.  The comment has been
removed.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I7a69b52a65dda815b533c605f581e22fc4445be0