LU-0000 lnet: always pass struct lnet_md by reference.
Both LNetMDAttach and LNetMDBind expected a struct lnet_md to be
passed by value.  This requires coping the data structure onto the
stack, which is a waste of stack space and brings no value.

So change the to expect a reference, and declare it 'const' to be sure
it doesn't get changed.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I343797d1e70cc85fde92d544e56536e982e02973