LU-11085 ldlm: simplify use of interval-tree.
The interval tree use for keeping track of extent locks is currently
separate from those locks themselves.  A separate 'ldlm_interval'
structure is allocated and linked to all locks which have the same
extent.

This requires that the interval tree library handles an insert when
exactly the same interval already exists differently from any other
insert.  No other users of the interval tree library wants this, and
the library which is part of linux doesn't support it.  So it would be
good to remove this requirement.

This patch changes the library, remove the 'ldlm_interval' structure,
and stores each lock in the tree.  This substantially simplifies a lot
of code, but has some costs.

The ldlm_lock is now larger - it contains three pointers for the
rbtree were previously it had one, and it now has an extra copy of the
range start/end.  These will be resolved in later patches but removing
duplication and sharing spare with other fields that aren't used for
extent locks.

The extent-tree can now be substantially larger as it now contains
every lock for a given extent rather than each extent only once.  The
the depth of the tree grows with the log of the number of element,
this isn't an enormous cost, but it may still be measurable.
In particular, locks that cover the full extent [0..MAX] are common
and can swamp other locks (citation needed).  Such locks can be easily
kept in a separate list.  This will restore some of the code
complexity, but is otherwise of little cost.

Linux-commit: 71236833ad7a98b69e6e675efefbdc04a74c1d4b

Change-Id: I6c82d971aabd02bb036ac0bd27a934d48e972895
Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: James Simmons <jsimmons@infradead.org>