LU-13903 build: Always include OpenSFS UAPI headers
Various utils code including util/parser.c and util/nidstrings.c
include files from the lnet,lustre/uapi directory, but that
is not included in CPPFLAGS when configured with --disable-modules.

There is no need to isolate the assignment to that case - we can
safely add these directories unconditionally.

Test-Parameters: trivial
Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I9974ad78ccabe84b1a0054c41bbc017503be231c