LU-0000 tests: pass -D flag to debugfs
debugfs normally accesses cached data for a block device.
The ldiskfs filesystem ignores that cache.
So if you:
 - use debugfs to examine a block device
 - make changes via the filesystem.
 - use debugfs again

The second debugfs call could access data that was cache from the
first call, and so might not see updates made to the filesystem.
This can cause intermitted test fails.

If we pass -D to debugfs, it uses O_DIRECT and always gets
current data.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: Ib13b409e169b9391de67cb567fc749b9da658162