LU-11085 ldlm: convert ldlm extent locks to linux extent-tree
As linux has a fully customizable extent tree implementation, use that
instead of the one in lustre.  This removes the need to store the
extent endpoints in the lock twice, thus recovering some of the space
wasted in a previous patch.

It also allows iteration loops to be in-line rather than requiring a
callback - though in some cases we keep the callback.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: Ie28c6fb0d40d2c92c7067c7a79f48ee1fc633ce9