LU-6142 llite: remove ll_dir_chain
ll_dir_chain is an empty structure that was introduced 13 years ago
and never used.  As such it is clutter.
Remove it an all references to it.  If the need which proposed it
ever does get addresses, having a clean slate in which to create
a suitable implementation is probably preferred.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I1eff80add8851ea11fff979f9b2d25916d6d3eed