LU-00000 ptlrpc: only build nrs_delay for server builds.
nrs_delay is only needed server-side, but it built
for clients too.  Other nrs_*.c files are built on for
client-only, but their content in #ifdef-ed away.

Move all these nrs_*.c files to a separate section of the Makefile
which is only built for server, and remove the #ifdefs.
This improves consistency and makes it easier to determine which
files are needed for the client.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: Ie25ed5ac837ff5f891824fad31ab258923840ec0