LU-00000 lnet: simplify cfs_parse_nidlist()
By duplicating the string being parsed, we can use mutating parsign
primitives and simplify the code.

Note that the kernel-space cfs_parse_nidlist() is now different from
the user-space version.  As they are both declared in the same header
file, this needs an #ifdef until the headers can be separated.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I1d5155a1cee82f798bec0863d80d75af92399cf1